Gave up on stack traces for snapshot names.
Switched from stack traces to relying on [CallerFilePath] and [CallerMemberName] to name snapshot directories and files, respectively. It's still possible to override the snapshot file, including using subdirectories, but the root will always be named for the source file from [CallerFilePath].

I could never reproduce it locally, but sometimes the stack traces seemed to not match my expectations so tests used the wrong snapshot file.